[Homework 11] add homework

+++ 11_SpringBoot/src/main/java/ru/otus/hw/dto/request/ProductCreateRequestDto.java
@@ -6,6 +6,8 @@ import lombok.Data;
 import lombok.NoArgsConstructor;
 import ru.otus.hw.model.Product;
 
+import java.math.BigDecimal;
+
 @Data
 @Builder
 @NoArgsConstructor
@@ -18,7 +20,7 @@ public class ProductCreateRequestDto {
     public static Product toEntity(ProductCreateRequestDto productCreateRequestDto) {
         Product product = new Product();
         product.setName(productCreateRequestDto.getName());
-        product.setPrice(productCreateRequestDto.getPrice());
+        product.setPrice(BigDecimal.valueOf(productCreateRequestDto.getPrice()));
         return product;
     }
